Restructure console layout
This change restructures the console layout to prevent memory corruption
and improve coordinate visibility. The console is relocated to the top of
the screen at y=0 and its horizontal margins are reduced to 4 pixels to
provide more space for long coordinate strings. A manual clearing loop is
added to the top 20 pixels of each frame buffer to prevent text smearing
between frames. Also the fractal rendering loop is adjusted to start at
row 20 as the text and fractal occupy distinct areas of the display.

Documentation for the PackYUVPair function was also updated to the use
of a palette pointer instead of an index.